+++ tests/input/test_factory.py
@@ -1,18 +1,18 @@
 # AutoTransform
 # Large scale, component based code modification library
 #
-# Licensed under the MIT License <http://opensource.org/licenses/MIT
+# Licensed under the MIT License <http://opensource.org/licenses/MIT>
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2022-present Nathan Rockenbach <http://github.com/nathro>
 
-"""Tests that the input factory has all types included."""
+"""Tests that the InputFactory has all types included."""
 
 from autotransform.input.factory import InputFactory
 from autotransform.input.type import InputType
 
 
 def test_all_enum_values_present():
-    """Ensures that all values from the enum are present in the factory getters."""
+    """Ensures that all values from the enum are present in the factory map."""
     missing_values = []
     for input_type in InputType:
         # pylint: disable=protected-access